+++ CHANGELOG.md
@@ -18,11 +18,14 @@ _Note: Gaps between patch versions are faulty, broken or test releases._
 * Fixed a bug when the `component` getter is `undefined`.
   Now the getter can return a promise in that case `base/b-dynamic-page`.
 
+* Fixed canceling of previous tasks with `scrollToProxy` `iPage`
 * Restored support of favicons `build/static.gulp` `iStaticPage`
 
 #### :house: Internal
 
-* Added tests `iStaticPage`
+* Added tests:
+  * `iPage`
+  * `iStaticPage`
 
 ## v3.0.0-rc.214 (2021-07-22)